requirements.txt
python-dotenv>=0.20.0
git+https://github.com/Rapptz/discord.py@master
